Add protein interaction input to PIS.
This file is rarely updated and was moved manually from release to release by the BE team. This file is now hosted in Google Storage with a timestamp. When the file needs to be updated the data team will do this, and place the new file in the relevant directly for PIS collection.